refactor(social): share monitor launchd helpers
