+++ scripts/bluesky-monitor-launchd.js
@@ -1,29 +1,25 @@
 #!/usr/bin/env node
 'use strict';
 
-const cp = require('node:child_process');
 const fs = require('node:fs');
 const os = require('node:os');
 const path = require('node:path');
 const { parseEnvFallback } = require('./social-analytics/load-env');
+const {
+  LAUNCHCTL_ENV,
+  buildMonitorPlist,
+  escapePlistString,
+  installLaunchAgent,
+  isCliInvocation: isLaunchdCliInvocation,
+  loadLaunchAgent,
+  parseArgs,
+  plistPathForLabel: defaultPlistPathForLabel,
+  runLaunchctl,
+} = require('./monitor-launchd-common');
 
 const DEFAULT_LABEL = 'com.thumbgate.bluesky-monitor';
 const DEFAULT_INTERVAL_MINUTES = 15;
 const DEFAULT_REPO_DIR = path.resolve(__dirname, '..');
-const LAUNCHCTL = '/bin/launchctl';
-const LAUNCHCTL_ENV = Object.freeze({
-  PATH: '/usr/bin:/bin:/usr/sbin:/sbin',
-});
-
-function escapePlistString(value) {
-  return String(value || '')
-    .replaceAll('&', '&amp;')
-    .replaceAll('<', '&lt;')
-    .replaceAll('>', '&gt;')
-    .replaceAll('"', '&quot;')
-    .replaceAll("'", '&#39;');
-}
-
 function buildBlueskyMonitorPlist(options = {}) {
   const label = options.label || DEFAULT_LABEL;
   const repoDir = path.resolve(options.repoDir || DEFAULT_REPO_DIR);
@@ -34,45 +30,24 @@ function buildBlueskyMonitorPlist(options = {}) {
   const publishApproved = options.publishApproved ? 'true' : 'false';
   const nodeBin = options.nodeBin || process.execPath;
 
-  return `<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-  <key>Label</key>
-  <string>${escapePlistString(label)}</string>
-  <key>ProgramArguments</key>
-  <array>
-    <string>/bin/bash</string>
-    <string>${escapePlistString(scriptPath)}</string>
-  </array>
-  <key>WorkingDirectory</key>
-  <string>${escapePlistString(repoDir)}</string>
-  <key>StartInterval</key>
-  <integer>${intervalSeconds}</integer>
-  <key>RunAtLoad</key>
-  <true/>
-  <key>StandardOutPath</key>
-  <string>${escapePlistString(path.join(logDir, 'bluesky-monitor-launchd.stdout.log'))}</string>
-  <key>StandardErrorPath</key>
-  <string>${escapePlistString(path.join(logDir, 'bluesky-monitor-launchd.stderr.log'))}</string>
-  <key>EnvironmentVariables</key>
-  <dict>
-    <key>HOME</key>
-    <string>${escapePlistString(os.homedir())}</string>
-    <key>THUMBGATE_REPO_DIR</key>
-    <string>${escapePlistString(repoDir)}</string>
-    <key>THUMBGATE_BLUESKY_PUBLISH_APPROVED</key>
-    <string>${escapePlistString(publishApproved)}</string>
-    <key>NODE_BIN</key>
-    <string>${escapePlistString(nodeBin)}</string>
-  </dict>
-</dict>
-</plist>
-`;
+  return buildMonitorPlist({
+    label,
+    repoDir,
+    intervalSeconds,
+    scriptPath,
+    stdoutPath: path.join(logDir, 'bluesky-monitor-launchd.stdout.log'),
+    stderrPath: path.join(logDir, 'bluesky-monitor-launchd.stderr.log'),
+    env: {
+      HOME: os.homedir(),
+      THUMBGATE_REPO_DIR: repoDir,
+      THUMBGATE_BLUESKY_PUBLISH_APPROVED: publishApproved,
+      NODE_BIN: nodeBin,
+    },
+  });
 }
 
 function plistPathForLabel(label = DEFAULT_LABEL) {
-  return path.join(os.homedir(), 'Library', 'LaunchAgents', `${label}.plist`);
+  return defaultPlistPathForLabel(label);
 }
 
 function loadEnvSnapshot({ repoDir = DEFAULT_REPO_DIR, env = process.env } = {}) {
@@ -177,66 +152,13 @@ function buildBlueskyMonitorStatus(options = {}) {
   };
 }
 
-function runLaunchctl(args, stdio) {
-  return cp.execFileSync(LAUNCHCTL, args, { env: LAUNCHCTL_ENV, stdio });
-}
-
-function loadLaunchAgent(plistPath) {
-  const uid = typeof process.getuid === 'function' ? process.getuid() : null;
-  if (uid !== null) {
-    try {
-      runLaunchctl(['bootout', `gui/${uid}`, plistPath], 'ignore');
-    } catch {
-      // Ignore if the agent was not loaded yet.
-    }
-    runLaunchctl(['bootstrap', `gui/${uid}`, plistPath], 'pipe');
-    runLaunchctl(['enable', `gui/${uid}/${path.basename(plistPath, '.plist')}`], 'pipe');
-    return;
-  }
-
-  try {
-    runLaunchctl(['unload', plistPath], 'ignore');
-  } catch {
-    // Ignore if the agent was not loaded yet.
-  }
-  runLaunchctl(['load', '-w', plistPath], 'pipe');
-}
-
 function installBlueskyMonitorLaunchAgent(options = {}) {
   const label = options.label || DEFAULT_LABEL;
   const repoDir = path.resolve(options.repoDir || DEFAULT_REPO_DIR);
   const plistPath = options.plistPath || plistPathForLabel(label);
   const plist = buildBlueskyMonitorPlist({ ...options, label, repoDir });
 
-  if (options.dryRun) return { ok: true, plistPath, plist, installed: false };
-
-  fs.mkdirSync(path.dirname(plistPath), { recursive: true });
-  fs.mkdirSync(path.join(repoDir, '.thumbgate'), { recursive: true });
-  fs.writeFileSync(plistPath, plist, 'utf8');
-  loadLaunchAgent(plistPath);
-  return { ok: true, plistPath, plist, installed: true };
-}
-
-function parseArgs(argv = process.argv.slice(2)) {
-  const args = { _: [] };
-  for (let index = 0; index < argv.length; index += 1) {
-    const token = argv[index];
-    if (!token.startsWith('--')) {
-      args._.push(token);
-      continue;
-    }
-    const option = token.slice(2);
-    const equalsIndex = option.indexOf('=');
-    if (equalsIndex >= 0) {
-      args[option.slice(0, equalsIndex)] = option.slice(equalsIndex + 1);
-    } else if (argv[index + 1] && !argv[index + 1].startsWith('--')) {
-      args[option] = argv[index + 1];
-      index += 1;
-    } else {
-      args[option] = true;
-    }
-  }
-  return args;
+  return installLaunchAgent({ options, label, repoDir, plistPath, plist });
 }
 
 function main(argv = process.argv.slice(2)) {
@@ -270,8 +192,7 @@ function main(argv = process.argv.slice(2)) {
 }
 
 function isCliInvocation(argv = process.argv) {
-  const invokedPath = argv[1];
-  return invokedPath ? path.resolve(invokedPath) === __filename : false;
+  return isLaunchdCliInvocation(__filename, argv);
 }
 
 if (isCliInvocation()) {
